Refactor Input
