+++ src/index.ts
@@ -95,6 +95,7 @@ export { default as Avatar } from './components/Avatar'
 export { default as Button } from './components/Button'
 export { default as CheckBox } from './components/Checkbox'
 export { default as FormField } from './components/FormField'
+export { default as Input } from './components/Input'
 export { default as Menu } from './components/Menu'
 export { default as MenuItem } from './components/MenuItem'
 export { default as RadioButton } from './components/RadioButton'
@@ -102,7 +103,6 @@ export { default as RepositoryCard } from './components/RepositoryCard'
 export { default as Sidebar } from './components/Sidebar'
 export { default as Switch } from './components/Switch'
 export { default as Tag } from './components/Tag'
-export { default as TextInput } from './components/TextInput'
 export { default as Token } from './components/Token'
 export { default as UserCard } from './components/UserCard'
 // Hooks
